+++ packages/server/bff-utils/src/client/generate-client.ts
@@ -1,3 +1,4 @@
+import * as path from 'path';
 import { HttpMethod } from '../constant';
 import { isAllowedHttpMethod } from '../utils';
 import { Result, Ok, Err } from './result';
@@ -20,6 +21,8 @@ export type GenClientOptions = {
 
 export const DEFAULT_CLIENT_REQUEST_CREATOR = '@modern-js/create-request';
 
+// eslint-disable-next-line eslint-comments/disable-enable-pair
+/* eslint-disable max-statements */
 export const generateClient = async ({
   resourcePath,
   source,
@@ -36,6 +39,17 @@ export const generateClient = async ({
     requestCreator = requireResolve(
       `${DEFAULT_CLIENT_REQUEST_CREATOR}${target ? `/${target}` : ''}`,
     ).replace(/\\/g, '/');
+  } else {
+    // 这里约束传入的 requestCreator 包也必须有两个导出 client 和 server，因为目前的机制 client 和 server 要导出不同的 configure 函数；该 api 不对使用者暴露，后续可优化
+    let resolvedPath = requestCreator;
+    try {
+      resolvedPath = path.dirname(requireResolve(requestCreator));
+    } catch (error) {}
+    // eslint-disable-next-line no-param-reassign
+    requestCreator = `${resolvedPath}${target ? `/${target}` : ''}`.replace(
+      /\\/g,
+      '/',
+    );
   }
 
   const routeRsult = getRouteName(resourcePath, apiDir);