+++ packages/toolkit/utils/src/constants.ts
@@ -119,7 +119,10 @@ export const INTERNAL_PLUGINS: {
   '@modern-js/plugin-static-hosting': {
     cli: '@modern-js/plugin-static-hosting/cli',
   },
-  '@modern-js/plugin-polyfill': { server: '@modern-js/plugin-polyfill' },
+  '@modern-js/plugin-polyfill': {
+    cli: '@modern-js/plugin-polyfill/cli',
+    server: '@modern-js/plugin-polyfill',
+  },
   '@modern-js/plugin-multiprocess': {
     cli: '@modern-js/plugin-multiprocess/cli',
   },
@@ -245,7 +248,7 @@ export const PLUGIN_SCHEMAS = {
     },
     {
       target: 'tools.jest',
-      schema: { typeof: ['object'] },
+      schema: { typeof: ['object', 'function'] },
     },
   ],
   '@modern-js/plugin-garfish': [