+++ packages/cli/babel-preset-module/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @modern-js/babel-preset-module
 
+## 1.3.0
+
+### Minor Changes
+
+- 0d9516f3: fix resolve static file and add -p params
+
+### Patch Changes
+
+- Updated dependencies [823809c6]
+  - @modern-js/utils@1.2.1
+
+## 1.2.0
+
+### Minor Changes
+
+- cfe11628: Make Modern.js self bootstraping
+
+### Patch Changes
+
+- 1ebc7ee2: fix: @babel/core version
+- Updated dependencies [2da09c69]
+- Updated dependencies [c3d46ee4]
+- Updated dependencies [cfe11628]
+- Updated dependencies [1ebc7ee2]
+  - @modern-js/utils@1.2.0
+  - @modern-js/babel-preset-lib@1.2.0
+  - @modern-js/babel-chain@1.2.0
+
 ## 1.1.4
 
 ### Patch Changes