+++ packages/cli/plugin-proxy/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @modern-js/plugin-proxy
 
+## 1.2.1
+
+### Patch Changes
+
+- e2d3a575: fix extending core config interface
+- e2d3a575: fix extending core config interface
+- Updated dependencies [823809c6]
+- Updated dependencies [4584cc04]
+- Updated dependencies [7c19fd94]
+  - @modern-js/utils@1.2.1
+  - @modern-js/core@1.3.1
+
+## 1.2.0
+
+### Minor Changes
+
+- cfe11628: Make Modern.js self bootstraping
+
+### Patch Changes
+
+- f6932de7: fix proxy bin error
+- bc87647d: fix: terminate whistle proxy when exiting the process
+- Updated dependencies [2da09c69]
+- Updated dependencies [fc71e36f]
+- Updated dependencies [c3d46ee4]
+- Updated dependencies [cfe11628]
+  - @modern-js/utils@1.2.0
+  - @modern-js/core@1.3.0
+
 ## 1.1.1
 
 ### Patch Changes