+++ packages/generator/generators/mwa-generator/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @modern-js/mwa-generator
 
+## 1.3.1
+
+### Patch Changes
+
+- 3c5621f0: fix tsconfig template
+
+## 1.3.0
+
+### Minor Changes
+
+- cfe11628: Make Modern.js self bootstraping
+
+### Patch Changes
+
+- 2c049918: feat: polyfill function support enable using new command
+
 ## 1.2.4
 
 ### Patch Changes